+++ Venmo/src/main/java/com/braintreepayments/api/VenmoAccountNonce.java
@@ -19,6 +19,9 @@ public class VenmoAccountNonce extends PaymentMethodNonce {
     private static final String VENMO_DETAILS_KEY = "details";
     private static final String VENMO_USERNAME_KEY = "username";
 
+    private static final String VENMO_PAYMENT_METHOD_ID_KEY = "paymentMethodId";
+    private static final String VENMO_PAYMENT_METHOD_USERNAME_KEY = "userName";
+    
     private final String username;
 
     static VenmoAccountNonce fromJSON(JSONObject inputJson) throws JSONException {
@@ -29,11 +32,21 @@ public class VenmoAccountNonce extends PaymentMethodNonce {
             json = inputJson;
         }
 
-        String nonce = json.getString(PAYMENT_METHOD_NONCE_KEY);
-        boolean isDefault = json.optBoolean(PAYMENT_METHOD_DEFAULT_KEY, false);
+        String nonce;
+        boolean isDefault;
+        String username;
 
-        JSONObject details = json.getJSONObject(VENMO_DETAILS_KEY);
-        String username = details.getString(VENMO_USERNAME_KEY);
+        if (json.has(VENMO_PAYMENT_METHOD_ID_KEY)) {
+            isDefault = false;
+            nonce = json.getString(VENMO_PAYMENT_METHOD_ID_KEY);
+            username = json.getString(VENMO_PAYMENT_METHOD_USERNAME_KEY);
+        } else {
+            nonce = json.getString(PAYMENT_METHOD_NONCE_KEY);
+            isDefault = json.optBoolean(PAYMENT_METHOD_DEFAULT_KEY, false);
+
+            JSONObject details = json.getJSONObject(VENMO_DETAILS_KEY);
+            username = details.getString(VENMO_USERNAME_KEY);
+        }
 
         return new VenmoAccountNonce(nonce, username, isDefault);
     }