+++ Venmo/src/main/java/com/braintreepayments/api/VenmoRequest.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import androidx.annotation.Nullable;
+
 /**
  A VenmoRequest specifies options that contribute to the Venmo flow
  */
@@ -7,6 +9,18 @@ public class VenmoRequest {
 
     private boolean shouldVault;
     private String profileId;
+    private String displayName;
+
+    private final @VenmoPaymentMethodUsage int paymentMethodUsage;
+
+    /**
+     * Request to tokenize a Venmo account.
+     *
+     * @param paymentMethodUsage {@link VenmoPaymentMethodUsage} for the tokenized Venmo account: either multi-use or single use.
+     */
+    public VenmoRequest(@VenmoPaymentMethodUsage int paymentMethodUsage) {
+        this.paymentMethodUsage = paymentMethodUsage;
+    }
 
     /**
      * @param shouldVault Optional - Whether or not to automatically vault the Venmo Account.
@@ -41,5 +55,37 @@ public class VenmoRequest {
     public String getProfileId() {
         return profileId;
     }
-}
 
+    /**
+     * @return {@link VenmoPaymentMethodUsage} for the tokenized Venmo account: either multi-use or single use.
+     */
+    public @VenmoPaymentMethodUsage int getPaymentMethodUsage() {
+        return paymentMethodUsage;
+    }
+
+    /**
+     * @param displayName Optional. The business name that will be displayed in the Venmo app payment approval screen. Only used by merchants onboarded as PayFast channel partners.
+     */
+    public void setDisplayName(@Nullable String displayName) {
+        this.displayName = displayName;
+    }
+
+    /**
+     * @return The business name that will be displayed in the Venmo app payment approval screen. Only used by merchants onboarded as PayFast channel partners.
+     */
+    @Nullable
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    String getPaymentMethodUsageAsString() {
+        switch (paymentMethodUsage) {
+            case VenmoPaymentMethodUsage.SINGLE_USE:
+                return "SINGLE_USE";
+            case VenmoPaymentMethodUsage.MULTI_USE:
+                return "MULTI_USE";
+            default:
+                return null;
+        }
+    }
+}